[refactor] 패키지명 컨벤션 변경에 의한 프로젝트 도메인과의 충돌 제어
