+++ module-domain/src/main/java/com/welcommu/moduledomain/project/ProjectsStatus.java
@@ -1,4 +1,4 @@
-package com.welcommu.moduledomain.projects;
+package com.welcommu.moduledomain.project;
 
 public enum ProjectsStatus {
     ACTIVE,
